backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
django-filter>=24.0
django-cors-headers>=4.3
psycopg[binary]>=3.1
gunicorn>=22.0
djangorestframework-simplejwt>=2.3
PyJWT>=2.8
